+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- Changed `alltoall` to return the received splits as a second return value if non-uniform splits are sent ([#2631](https://github.com/horovod/horovod/pull/2631))
+
 ### Deprecated
 
 ### Removed